Revert "Add flags to input"
This reverts commit 333c78a6705077747459bab6e05a41471d525b06.
